==> experiments/6_latent_representation_visualstimuli.ipynb <==
import torch
import torch.nn as nn
import torch.nn.functional as F
import numpy as np
import matplotlib.pyplot as plt
import pickle
import os
from mne.datasets import visual_92_categories
import pandas as pd
import mne

with open('data/visual_stimuli/visual_stimuli_ica_preprocessed.pickle', 'rb') as f:
    data = pickle.load(f)

data.shape

list(np.linspace(0, 92, 3, dtype=int)[:2])

print(data.mean(), data.std())

# normalize data between 0 and 1
data = torch.Tensor((data - np.min(data)) / (np.max(data) - np.min(data)))

print(data.mean(), data.std())

data.shape

# save the labeled target class for each sample

data_path = visual_92_categories.data_path()

# Define stimulus - trigger mapping
fname = data_path / 'visual_stimuli.csv'
conds = pd.read_csv(fname)

conds_tensor = torch.Tensor(np.array(conds[['human', 'face', 'animal', 'natural']]))

conds

import pickle

with open('./data/visual_stimuli/events_label.pkl', 'rb') as f:
    labels = pickle.load(f)

target_class = torch.zeros(labels.shape[0], 4, dtype=torch.long)

for i, label in enumerate(labels):
    target_class[i] = conds_tensor[int(label.item())]

target_class.shape

import pickle

with open('./data/visual_stimuli/target_class.pickle', 'wb') as f:
    pickle.dump(target_class, f, protocol=pickle.HIGHEST_PROTOCOL)


# split data into train and test
s_train = int(number_samples*0.8)
s_test = number_samples - s_train

s_test

s_train

Load and train the transformer encoder model to perform classifiction of the visual sitmuli given the eeg/meg preprocess data responses.

import torch
import torch.nn as nn
import torch.nn.functional as F
import numpy as np
import matplotlib.pyplot as plt
import pickle
import os
from mne.datasets import visual_92_categories
import pandas as pd

with open('./data/visual_stimuli/target_class.pickle', 'rb') as f:
    target_class = pickle.load(f)
# with open('./data/visual_stimuli/events_label.pkl', 'rb') as f:
#     target_class = pickle.load(f)
with open('./data/visual_stimuli/visual_stimuli_ica_preprocessed.pickle', 'rb') as f:
    data_exp = pickle.load(f)

# normalize data between 0 and 1
data_exp = torch.Tensor((data_exp - np.min(data_exp)) / (np.max(data_exp) - np.min(data_exp)))

print(data_exp.mean(), data_exp.std())

# prepare data for training
number_samples = 2758

# add cls token
cls_ = torch.zeros(number_samples, 1, data_exp[0].shape[1])

data_exp = torch.cat((cls_, data_exp), dim=1)

# split data into train and test
s_train = int(number_samples*0.8)
s_test = number_samples - s_train

batch = 64

generator = torch.Generator().manual_seed(42)

data = torch.utils.data.TensorDataset(data_exp, target_class.type(torch.LongTensor))
training_samples, validation_samples = torch.utils.data.random_split(data, [s_train, s_test], generator=generator)

train = torch.utils.data.DataLoader(training_samples, batch_size=batch, shuffle=True)
val = torch.utils.data.DataLoader(validation_samples, batch_size=batch, shuffle=False)

from module._module import *

net = Net(dim_emb=601, dim_emb_transformer=64,number_classes=4, nhead=2, dim_feedforward=32).cuda()

net(data_exp[:64].cuda()).shape

from tranformer import training
from torch import optim

training(net, train, optimizer = optim.AdamW(net.parameters(), lr=0.001), batch_size=batch, epochs=100)

from tranformer import Netv2

net = Netv2(dim_emb=601, dim_emb_transformer=295,number_classes=4, nhead=1, dim_feedforward=16).cuda()

from tranformer import training
from torch import optim

training(net, train, optimizer = optim.AdamW(net.parameters(), lr=0.001), batch_size=batch, epochs=10)

from tranformer import test

test(net, val, batch_size=batch)


outputs = torch.nn.Softmax()(torch.randn((10,4)))
label = torch.randint(0, 2, (10,4))

(outputs == label).sum().item()/(10*4)

out = net(data_exp[:200].cuda())

class MLP(nn.Module):
    def __init__(self, dim_emb, dim_feedforward: int = 10, number_classes: int = 92, dropout: float =0.1, activation = nn.GELU):
        super(MLP, self).__init__()
        
        # Implementation of Feedforward model
        self.linear1 = nn.Linear(dim_emb, dim_feedforward)
        self.linear2 = nn.Linear(dim_feedforward, number_classes)
        
        self.dropout = nn.Dropout(dropout)
        self.activation = activation()

    def forward(self, x):
        
        x = self.linear2(self.dropout(self.activation(self.linear1(x))))    
            
        return x
    

nn.Linear((601**2), 10)(torch.randn((64,601,601)).reshape(64,601**2)).shape

out = torch.randn((295,1))


nn.AdaptiveMaxPool1d(out.shape[0]//4)(out.permute(1,0)).shape

nn.Conv2d(in_channels=3, out_channels=3, kernel_size=3, groups=3)

nn.Conv2d(in_channels=607, out_channels=1, kernel_size=(307,1), stride=1, dilation=1)(
    torch.randn((64,307,607)).permute(2,1,0)
).squeeze(0).permute(1,0).shape

torch.randn((64,307,607)).permute(2,1,0).shape
