==> experiments/9_face_processing_read_preprocess_data.ipynb <==
import glob
import mne
import numpy as np
import pandas as pd
import torch
import pickle



# list files
files_name_eeglab = sorted(glob.glob('./face_processing_data/*/eeg/sub-*_task-FaceRecognition_eeg.set'))
files_name_tsv = sorted(glob.glob('./face_processing_data/sub-*/eeg/sub-*_task-FaceRecognition_events.tsv'))



###############################################################

def get_onset_duration(df):
    """ This function gets the onset and duration given the df """
    onset = []
    duration = []
    trial_type = []
    label = []

    for i in range(len(df)):
        if df['event_type'][i] == 'button_press':
            if df['event_type'][i-1] == 'faces':
                onset.append(df['onset'][i-1])
                duration.append(df['onset'][i] - df['onset'][i-1])
                trial_type.append(df['trial_type'][i-1])
                label.append(df['value'][i-1])

                # code to get the max duration
                # this allows me to get a size 
                # of the window without overlaps
                # of 1.4 seconds
                # try:
                #     duration_max.append(df['onset'][i+1] - df['onset'][i-1])
                # except:
                #     pass


    return np.array(onset), np.array(duration), np.array(trial_type), np.array(label) # , np.array(duration_max)

###############################################################

def get_data_eeglab(files_name_eeglab, files_name_tsv, base_path):
    """ This function loads the data from the eeglab files and the tsv 
    files and saves them in a npz file.

    @param files_name_eeglab: list of strings
        list of the eeglab files
    @param files_name_tsv: list of strings
        list of the tsv files
    @param base_path: string
        path where to save the data

    @return: None
    """
    files_name = zip(files_name_eeglab, files_name_tsv)

    # target labels
    target_labels = torch.tensor([])
    target_trial_type = np.array([])
    X = torch.tensor([])
    event_dict_all = []
    subject_id = torch.tensor([])

    data = []
    for eeglab_file, tsv in files_name:

        #######################################################
        ######## LOAD EEG DATA FROM THE EEG FILE ##############
        #######################################################
        # load the data
        raw = mne.io.read_raw_eeglab(eeglab_file, preload=True)


        #######################################################
        ######## PERFORM ICA ##################################
        #######################################################

        # perform ica and plot
        # ica = mne.preprocessing.ICA(n_components=0.95, random_state=97, max_iter='auto')
        # ica.fit(raw)

        # ica_raw = ica.get_components()

        #######################################################
        ######## LOAD ANNOTATIONS FROM THE TSV FILE ###########
        #######################################################

        df = pd.read_csv(tsv, sep='\t')

        # get the onset and duration
        onset, duration, trial_type, label = get_onset_duration(df)

        # duration_max = np.max(duration) # 1.4s

        # get the target labels
        target_labels = torch.cat([target_labels, torch.from_numpy(label)])
        target_trial_type = np.concatenate([target_trial_type, trial_type])

        # add the annotations to the raw data
        annot = mne.Annotations(
            onset=list(onset),  # in seconds
            duration=[1.4]*onset.shape[0],  # in seconds, too
            description=list(trial_type),
            # id_code=list(df['value']),
        )

        # print the annotations
        # print(annot)

        raw2 = raw.copy().set_annotations(annot)

        # raw2.plot()
        # print(pd.DataFrame(raw2.annotations))

        # get events from the annotations
        events_from_annot, event_dict = mne.events_from_annotations(raw2)

        # epochs
        epochs = mne.Epochs(raw2, events_from_annot, event_id=event_dict, tmin=-0.2, tmax=0.55, preload=True)

        # filter the data
        epochs.filter(4, 80)

        # apply baseline correction
        epochs.apply_baseline((None, 0))

        # # plot the epochs
        # epochs.plot(n_epochs=10, n_channels=10)

        # subject id
        subject_id = torch.cat([subject_id,torch.tensor([int(i) for i in [eeglab_file.split('/')[2].split('-')[1]]])])

        # concatenate the data
        X = torch.cat([X, torch.from_numpy(epochs.get_data()).to(torch.float32)])

        # concatenate the event dict
        event_dict_all.append(event_dict)

        # Save the data
        data.append((eeglab_file.split('/')[2], epochs.get_data(), epochs.info, epochs.events, epochs.event_id, event_dict, epochs.tmin, epochs.tmax))

        for d in data:
            print('SAVED: '+base_path + 'data_{}.npz'.format(d[0]))
            np.savez_compressed(base_path + 'data_{}.npz'.format(d[0]), data=np.array(d[1],dtype='float32'))

        # # save the data
        torch.save(X, base_path + '/dataset/X.pt')
        torch.save(target_labels, base_path + '/dataset/target_labels.pt')
        # torch.save(target_trial_type, base_path + '/dataset/target_trial_type.pt')
        np.save(base_path + '/dataset/target_trial_type.npy', target_trial_type)
        # torch.save(subject_id, base_path + '/dataset/subject_id.pt')
        with open(base_path + '/dataset/event_dict_all.pkl', 'wb') as f:
            pickle.dump(event_dict_all, f)
        

base_path = './data/face_recognition_preprocessed_data/'
get_data_eeglab(files_name_eeglab, files_name_tsv, base_path)

# import pickle
# with open('./data/face_recognition_preprocessed_data/face_recognition_preprocessed_data.pkl', 'wb') as f:
#     pickle.dump(data, f)


import pandas as pd
import numpy as np


df = pd.read_csv('./face_processing_data/sub-002/eeg/sub-002_task-FaceRecognition_events.tsv', sep='\t')

def get_onset_duration(df):
    """ This function gets the onset and duration given the df """
    onset = []
    duration = []
    trial_type = []
    label = []

    for i in range(len(df)):
        if df['event_type'][i] == 'button_press':
            if df['event_type'][i-1] == 'faces':
                onset.append(df['onset'][i-1])
                duration.append(df['onset'][i] - df['onset'][i-1])
                trial_type.append(df['trial_type'][i-1])
                label.append(df['value'][i-1])

                # code to get the max duration
                # this allows me to get a size 
                # of the window without overlaps
                # of 1.4 seconds
                # try:
                #     duration_max.append(df['onset'][i+1] - df['onset'][i-1])
                # except:
                #     pass


    return np.array(onset), np.array(duration), np.array(trial_type), np.array(label) # , np.array(duration_max)

onset, duration, trial_type, label = get_onset_duration(df)

print(duration.max(), duration.mean(), duration.min()) # 1.4, 0.7, 0.2 # we will use 1.419s

torch.tensor([int(i) for i in [files_name_eeglab[0].split('/')[2].split('-')[1]]*10])

df = pd.read_csv('./face_processing_data/sub-002/eeg/sub-002_task-FaceRecognition_events.tsv', sep='\t')

df


import torch

base_path = './data/face_recognition_preprocessed_data/'
X = torch.load(base_path + '/dataset/X.pt')

X.shape


import glob
import mne
import numpy as np
import pandas as pd
import torch
import pickle



# list files
files_name_eeglab = sorted(glob.glob('./face_processing_data/*/eeg/sub-*_task-FaceRecognition_eeg.set'))
files_name_tsv = sorted(glob.glob('./face_processing_data/sub-*/eeg/sub-*_task-FaceRecognition_events.tsv'))



def get_onset_duration(df):
    """ This function gets the onset and duration given the df """
    onset = []
    duration = []
    trial_type = []
    label = []

    for i in range(len(df)):
        if df['event_type'][i] == 'button_press':
            if df['event_type'][i-1] == 'faces':
                onset.append(df['onset'][i-1])
                duration.append(df['onset'][i] - df['onset'][i-1])
                trial_type.append(df['trial_type'][i-1])
                label.append(df['value'][i-1])

                # code to get the max duration
                # this allows me to get a size 
                # of the window without overlaps
                # of 1.4 seconds
                # try:
                #     duration_max.append(df['onset'][i+1] - df['onset'][i-1])
                # except:
                #     pass


    return np.array(onset), np.array(duration), np.array(trial_type), np.array(label) # , np.array(duration_max)


def get_data_eeglab(files_name_eeglab, files_name_tsv, base_path):
    """ 
    """
    files_name = zip(files_name_eeglab, files_name_tsv)

    for eeglab_file, tsv in files_name:

        #######################################################
        ######## LOAD EEG DATA FROM THE EEG FILE ##############
        #######################################################
        # load the data
        raw = mne.io.read_raw_eeglab(eeglab_file, preload=True)

        df = pd.read_csv(tsv, sep='\t')

        # get the onset and duration
        onset, duration, trial_type, label = get_onset_duration(df)

        # duration_max = np.max(duration) # 1.4s


        # add the annotations to the raw data
        annot = mne.Annotations(
            onset=list(onset),  # in seconds
            duration=[1.4]*onset.shape[0],  # in seconds, too
            description=list(trial_type),
            # id_code=list(df['value']),
        )

        # print the annotations
        # print(annot)

        raw2 = raw.copy().set_annotations(annot)

        # raw2.plot()
        # print(pd.DataFrame(raw2.annotations))

        # get events from the annotations
        events_from_annot, event_dict = mne.events_from_annotations(raw2)

        # epochs
        epochs = mne.Epochs(raw2, events_from_annot, event_id=event_dict, tmin=-0.2, tmax=0.55, preload=True)

        # filter the data
        epochs.filter(4, 80)

        # apply baseline correction
        epochs.apply_baseline((None, 0))

        print(raw.info)
        print(epochs.info)
        print(epochs.times)
        break



base_path = './data/face_recognition_preprocessed_data/'
get_data_eeglab(files_name_eeglab, files_name_tsv, base_path)
