==> experiments/12_preprocess_emotion_recognition_data.ipynb <==
import glob
import numpy as np


file_name = sorted(glob.glob('./data/emotion*/*.npz'))
byte = 0
for file in file_name:
    print(file)
    data = np.load(file)
    print(data['data'].shape)
    # pad the secodn dimension
    data = np.pad(data['data'], ((0, 0), (0, 235 - data['data'].shape[1]), (0, 0)), 'constant', constant_values=0)
    print(data.shape)
    # dimension in bytes
    byte = byte + data.nbytes
    print(byte)

# full data size 11.3 GB

import glob
import numpy as np


file_name = sorted(glob.glob('./data/emotion*/*.npz'))
byte = 0
for file in file_name:
    print(file)
    data = np.load(file)
    print(data['data'][:,:,:12800].shape)
    # pad the secodn dimension
    data = np.pad(data['data'][:,:,:12800], ((0, 0), (0, 235 - data['data'].shape[1]), (0, 0)), 'constant', constant_values=0)
    print(data.shape)
    # dimension in bytes
    byte = byte + data.nbytes
    print(byte)

# full data size 5.55 GB

# preprocess the data
'./data/emotion_recognition_preprocessed_data/preprocessed/'


file_name = sorted(glob.glob('./data/emotion*/*.npz'))
participants = pd.read_csv('./emotion_recognition_data/participants.tsv', sep='\t').sort_values(by=['participant_id']).reset_index(drop=True)


byte = 0
for file in file_name:
    print(file)
    data = np.load(file)
    print(data['data'][:,:,:16800].shape)
    # pad the secodn dimension
    data = np.pad(data['data'][:,:,:16800], ((0, 0), (0, 235 - data['data'].shape[1]), (0, 0)), 'constant', constant_values=0)
    print(data.shape)
    # dimension in bytes
    byte = byte + data.nbytes
    print(byte)

import pandas as pd

# sort by participant_id

import torch

# torch.tensor(data).dtype
import numpy as np

import glob
import numpy as np

import pandas as pd


gender_map = {'M': 0, 'F': 1}
age_map = {'25-35': 0, '35-45': 1, '45-55': 2, '55+': 3}
part_id = {
    'sub-01': 0,
    'sub-02': 1,
    'sub-03': 2,
    'sub-04': 3,
    'sub-05': 4,
    'sub-06': 5,
    'sub-07': 6,
    'sub-08': 7,
    'sub-09': 8,
    'sub-10': 9,
    'sub-11': 10,
    'sub-12': 11,
    'sub-13': 12,
    'sub-14': 13,
    'sub-15': 14,
    'sub-16': 15,
    'sub-17': 16,
    'sub-18': 17,
    'sub-19': 18,
    'sub-20': 19,
    'sub-21': 20,
    'sub-22': 21,
    'sub-23': 22,
    'sub-24': 23,
    'sub-25': 24,
    'sub-26': 25,
    'sub-27': 26,
    'sub-28': 27,
    'sub-29': 28,
    'sub-30': 29,
    'sub-31': 30,
    'sub-32': 31,
    'sub-33': 32,
    'sub-34': 33,
    'sub-35': 34
}

emotions = {'awe': 0,
            'fear': 1,
            'joy': 2,
            'jealousy': 3,
            'excite': 4,
            'disgust': 5,
            'love': 6,
            'grief': 7,
            'content': 8,
            'frustration': 9,
            'happy': 10,
            'anger': 11,
            'relief': 12,
            'sad': 13,
            'compassion': 14}


info = torch.tensor([])
data_X = torch.tensor([])
data_y = torch.tensor([])

number_steps = 15
length_s = 16800
shift = 1000000000

ind = np.linspace(0, length_s, number_steps, dtype=int)
file_name = sorted(glob.glob('./data/emotion*/*.npz'))
participants = pd.read_csv('./emotion_recognition_data/participants.tsv', sep='\t').sort_values(by=['participant_id']).reset_index(drop=True)


byte = 0
# for file in file_name:
for file in file_name:
    print(file)
    data = np.load(file)
    data = np.pad(data['data'][:,:,:length_s] * shift, ((0, 0), (0, 235 - data['data'].shape[1]), (0, 0)), 'constant', constant_values=0)

    data_cutted = torch.cat([torch.stack([torch.tensor(d[:,m:M], dtype=torch.float32) for m, M in zip(ind[:-1], ind[1:])]) for d in data])

    participant_info = participants[participants['participant_id'] == file.split('/')[-1].split('_')[1].split('.')[0]].values[0]

    participant_info = torch.tensor(
                            np.array([np.array([part_id[participant_info[0]], gender_map[participant_info[1]], participant_info[2], participant_info[4]])]*data_cutted.shape[0]), 
                                dtype=torch.int16)

    subj_id = file.split('/')[-1].split('_')[1].split('.')[0]
    l = torch.tensor([emotions[emotion] for emotion in pd.read_csv(f'./data/emotion_recognition_preprocessed_data/data_{subj_id}.csv')['value'].values]*(number_steps-1))

    # save info in list
    info = torch.cat([info, participant_info])
    data_X = torch.cat([data_X, data_cutted])
    data_y = torch.cat([data_y, l])


    

# del
del data_cutted, participant_info, subj_id, l, data

# save data
torch.save(data_X, './data/emotion_recognition_preprocessed_data/preprocessed_cutted/data.pt')

del data_X

torch.save(info, './data/emotion_recognition_preprocessed_data/preprocessed_cutted/info.pt')
torch.save(data_y, './data/emotion_recognition_preprocessed_data/preprocessed_cutted/labels.pt')

import gc
gc.collect()

import torch

# torch.tensor(data).dtype
import numpy as np

import glob
import numpy as np

import pandas as pd


gender_map = {'M': 0, 'F': 1}
age_map = {'25-35': 0, '35-45': 1, '45-55': 2, '55+': 3}
part_id = {
    'sub-01': 0,
    'sub-02': 1,
    'sub-03': 2,
    'sub-04': 3,
    'sub-05': 4,
    'sub-06': 5,
    'sub-07': 6,
    'sub-08': 7,
    'sub-09': 8,
    'sub-10': 9,
    'sub-11': 10,
    'sub-12': 11,
    'sub-13': 12,
    'sub-14': 13,
    'sub-15': 14,
    'sub-16': 15,
    'sub-17': 16,
    'sub-18': 17,
    'sub-19': 18,
    'sub-20': 19,
    'sub-21': 20,
    'sub-22': 21,
    'sub-23': 22,
    'sub-24': 23,
    'sub-25': 24,
    'sub-26': 25,
    'sub-27': 26,
    'sub-28': 27,
    'sub-29': 28,
    'sub-30': 29,
    'sub-31': 30,
    'sub-32': 31,
    'sub-33': 32,
    'sub-34': 33,
    'sub-35': 34
}

emotions = {'awe': 0,
            'fear': 1,
            'joy': 2,
            'jealousy': 3,
            'excite': 4,
            'disgust': 5,
            'love': 6,
            'grief': 7,
            'content': 8,
            'frustration': 9,
            'happy': 10,
            'anger': 11,
            'relief': 12,
            'sad': 13,
            'compassion': 14}


info = torch.tensor([])
data_X = torch.tensor([])
data_y = torch.tensor([])

number_steps = 15
length_s = 16800
shift = 1000000000

ind = np.linspace(0, length_s, number_steps, dtype=int)
file_name = sorted(glob.glob('./data/emotion*/*.npz'))
participants = pd.read_csv('./emotion_recognition_data/participants.tsv', sep='\t').sort_values(by=['participant_id']).reset_index(drop=True)


byte = 0
# for file in file_name:
for file in file_name:
    print(file)
    data = np.load(file)
    # add shift
    data = np.pad(data['data'][:,:,:length_s] * shift, ((0, 0), (0, 235 - data['data'].shape[1]), (0, 0)), 'constant', constant_values=0)

    data_cutted = torch.tensor(data, dtype=torch.float32)

    participant_info = participants[participants['participant_id'] == file.split('/')[-1].split('_')[1].split('.')[0]].values[0]

    participant_info = torch.tensor(
                            np.array([np.array([part_id[participant_info[0]], gender_map[participant_info[1]], participant_info[2], participant_info[4]])]*data_cutted.shape[0]), 
                                dtype=torch.int16)

    subj_id = file.split('/')[-1].split('_')[1].split('.')[0]
    l = torch.tensor([emotions[emotion] for emotion in pd.read_csv(f'./data/emotion_recognition_preprocessed_data/data_{subj_id}.csv')['value'].values], dtype=torch.int16)

    # save info in list
    info = torch.cat([info, participant_info])
    data_X = torch.cat([data_X, data_cutted])
    data_y = torch.cat([data_y, l])


    

del data, data_cutted, participant_info, subj_id, l
# save tensor

torch.save(data_X, './data/emotion_recognition_preprocessed_data/preprocessed/data.pt')

del data_X
torch.save(info, './data/emotion_recognition_preprocessed_data/preprocessed/info.pt')
torch.save(data_y, './data/emotion_recognition_preprocessed_data/preprocessed/labels.pt')

import torch

data = torch.load('./data/emotion_recognition_preprocessed_data/preprocessed/data.pt')
info = torch.load('./data/emotion_recognition_preprocessed_data/preprocessed/info.pt')
labels = torch.load('./data/emotion_recognition_preprocessed_data/preprocessed/labels.pt')


from torch.utils.data import DataLoader, TensorDataset
from torch.utils.data.dataset import random_split

BACH_SIZE = 64

dataset = TensorDataset(data, info, labels)

train_size = int(0.8 * len(dataset))
val_size = int(0.1 * len(dataset))
test_size = len(dataset) - train_size - val_size

train_dataset, val_dataset, test_dataset = random_split(dataset, [train_size, val_size, test_size], generator=torch.Generator().manual_seed(42))

train_loader = DataLoader(train_dataset, batch_size=BACH_SIZE, shuffle=True)
val_loader = DataLoader(val_dataset, batch_size=BACH_SIZE, shuffle=True)
test_loader = DataLoader(test_dataset, batch_size=BACH_SIZE, shuffle=True)


np.load(file)['data'].mean()
